pyproject.toml
# Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
#
# NVIDIA CORPORATION, its affiliates and licensors retain all intellectual
# property and proprietary rights in and to this material, related
# documentation and any modifications thereto. Any use, reproduction,
# disclosure or distribution of this material and related documentation
# without an express license agreement from NVIDIA CORPORATION or
# its affiliates is strictly prohibited.

[project]
name = "cuTAMP"
version = "0.0.1"
description = "cuTAMP"
readme = "README.md"
requires-python = ">=3.10"
keywords = ["robotics"]
authors = []
classifiers = [
    "Programming Language :: Python :: 3",
]

dependencies = [
    # Core
    "einops",
    "jaxtyping",
    "numpy",
    "omegaconf",
    "open3d",
    "pybullet",
    "rerun-sdk",
    "roma",
    "tqdm",
    "trimesh",
    "yourdfpy",
    "torch==2.7.0",
    "torchaudio==2.7.0",
    "torchvision==0.22.0",
    # VLM stacks
    "flash-linear-attention>=0.4.2",
    "transformers==5.4.0",
    # Third-party integrations
    "causal-conv1d",
    "nvidia-curobo",
]

[project.optional-dependencies]
dev = [
    "ruff",
]

[tool.ruff]
line-length = 120
exclude = ["cutamp/robots/assets"]

[project.scripts]
cutamp-demo = "cutamp.scripts.run_cutamp:entrypoint"
cutamp-collect-retrieval = "cutamp.scripts.collect_retrieval_data:entrypoint"
cutamp-benchmark-retrieval = "cutamp.scripts.benchmark_retrieval:entrypoint"
cutamp-vlm-demo = "cutamp.scripts.run_vlm_tamp:entrypoint"
cutamp-benchmark-vlm = "cutamp.scripts.benchmark_vlm:entrypoint"

[tool.setuptools.packages]
find = { include = ["cutamp", "cutamp.*"] }

[tool.uv]
no-build-isolation-package = [
    "causal-conv1d",
]

[tool.uv.sources]
torchaudio = { index = "torch" }
torch = { index = "torch" }
torchvision = { index = "torch" }

# Git sources
causal-conv1d = { git = "https://github.com/Dao-AILab/causal-conv1d.git" }
nvidia-curobo = { git = "https://github.com/NVlabs/curobo.git" }

[[tool.uv.index]]
explicit = true
name = "torch"
url = "https://download.pytorch.org/whl/cu128"

[[tool.uv.index]]
explicit = false
name = "pypi"
url = "https://pypi.org/simple"

[tool.uv.extra-build-dependencies]
nvidia-curobo = ["torch"]
